+++ NEWS.md
@@ -1,3 +1,6 @@
+# Version UNRELEASED
+
+
 # Version 22.9.5
 
 * Improved verse lookup regular expressions